requirements.txt
python-dotenv
streamlit
tensorflow
transformers
google.generativeai
spacy